requirements.txt
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
gcloud==0.18.3
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy
oauth2client==4.1.3
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.0
pandas==2.2.3
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rsa==4.9
six==1.17.0
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn